// src/shared/types.d.ts
interface CallToAction {
  text: string;
  href: string;
  icon?: Function;
  onClick?: Function;
  targetBlank?: boolean;
  btnText?: 'uppercase' | 'capitalize';
  btnType?: 'primary' | 'secondary';
}